rename types.

+++ src/main/kotlin/com/github/aiczk/syncweather/command/json/MaxTemp.kt
@@ -11,29 +11,29 @@ data class MaxTemp(
     @SerialName("hour")
     val hour: Int?,
     @SerialName("is_yearly_max_temp")
-    val isYearlyMaxTemp: Int?,
+    val isYearlyMaxTemp: Double?,
     @SerialName("max_temp_record")
     val maxTempRecord: Double?,
     @SerialName("max_temp_record_day")
-    val maxTempRecordDay: Int?,
+    val maxTempRecordDay: Double?,
     @SerialName("max_temp_record_month")
-    val maxTempRecordMonth: Int?,
+    val maxTempRecordMonth: Double?,
     @SerialName("max_temp_record_q")
     val maxTempRecordQ: Int?,
     @SerialName("max_temp_record_same_month")
     val maxTempRecordSameMonth: Double?,
     @SerialName("max_temp_record_same_month_day")
-    val maxTempRecordSameMonthDay: Int?,
+    val maxTempRecordSameMonthDay: Double?,
     @SerialName("max_temp_record_same_month_month")
-    val maxTempRecordSameMonthMonth: Int?,
+    val maxTempRecordSameMonthMonth: Double?,
     @SerialName("max_temp_record_same_month_q")
     val maxTempRecordSameMonthQ: Int?,
     @SerialName("max_temp_record_same_month_year")
-    val maxTempRecordSameMonthYear: Int?,
+    val maxTempRecordSameMonthYear: Double?,
     @SerialName("max_temp_record_year")
-    val maxTempRecordYear: Int?,
+    val maxTempRecordYear: Double?,
     @SerialName("max_temp_start_year")
-    val maxTempStartYear: Int?,
+    val maxTempStartYear: Double?,
     @SerialName("minute")
     val minute: Int?,
     @SerialName("month")
@@ -41,23 +41,23 @@ data class MaxTemp(
     @SerialName("temp_daily_max")
     val tempDailyMax: Double?,
     @SerialName("temp_daily_max_corresponding_month")
-    val tempDailyMaxCorrespondingMonth: Int?,
+    val tempDailyMaxCorrespondingMonth: Double?,
     @SerialName("temp_daily_max_corresponding_season_in_month")
-    val tempDailyMaxCorrespondingSeasonInMonth: Int?,
+    val tempDailyMaxCorrespondingSeasonInMonth: Double?,
     @SerialName("temp_daily_max_diff_to_avg")
     val tempDailyMaxDiffToAvg: Double?,
     @SerialName("temp_daily_max_diff_to_yesterday")
-    val tempDailyMaxDiffToYesterday: Int?,
+    val tempDailyMaxDiffToYesterday: Double?,
     @SerialName("temp_daily_max_new_record_in_decade")
-    val tempDailyMaxNewRecordInDecade: Int?,
+    val tempDailyMaxNewRecordInDecade: Double?,
     @SerialName("temp_daily_max_new_record_month")
-    val tempDailyMaxNewRecordMonth: Int?,
+    val tempDailyMaxNewRecordMonth: Double?,
     @SerialName("temp_daily_max_q")
     val tempDailyMaxQ: Int?,
     @SerialName("temp_daily_max_time_hour")
-    val tempDailyMaxTimeHour: Int?,
+    val tempDailyMaxTimeHour: Double?,
     @SerialName("temp_daily_max_time_minute")
-    val tempDailyMaxTimeMinute: Int?,
+    val tempDailyMaxTimeMinute: Double?,
     @SerialName("temp_daily_max_time_q")
     val tempDailyMaxTimeQ: Int?,
     @SerialName("updatedAt")
@@ -67,11 +67,11 @@ data class MaxTemp(
     @SerialName("yearly_max_temp")
     val yearlyMaxTemp: Double?,
     @SerialName("yearly_max_temp_day")
-    val yearlyMaxTempDay: Int?,
+    val yearlyMaxTempDay: Double?,
     @SerialName("yearly_max_temp_month")
-    val yearlyMaxTempMonth: Int?,
+    val yearlyMaxTempMonth: Double?,
     @SerialName("yearly_max_temp_q")
     val yearlyMaxTempQ: Int?,
     @SerialName("yearly_max_temp_year")
-    val yearlyMaxTempYear: Int?
+    val yearlyMaxTempYear: Double?
 )
